Week 12
The Netflix recommandation algo was harder than tought to implement. I need to remember to first make a plan of all the methods and class then write them. Choosing what object to use is important and the consequence can make debugging harder. @Todo Make the code cleaner as an exercice